moves utils files out of page directory changes function delivery to react components adjusts actice <> isActive naming for states
