+++ next/pages/guidelines/color/usage/background.tsx
@@ -9,7 +9,7 @@ import { H2, H3, LI, P, UL } from '../../../../components/mdx/components';
 import ExampleBox from '../../../../components/example-box';
 import UsageCategory from './usage-categories';
 import ColorUsageNav from './color-usage-nav';
-import { Usage, Image } from '../utils';
+import { Usage, Image } from '../../../../utils/color-usage-types';
 
 import intro from '../../../../images/pages/guide/product/color/usage/background/intro.png';
 import neutral from '../../../../images/pages/guide/product/color/usage/background/neutral.png';
